Introduce ID field for environment mappings
We were updating env slugs on the fly - this causes issues when terraform tries to re-calcualate the plan for change.
There is no clear or nice way to remove the plan because SDK calls DiffSuppressFunc on individual values not on whole plan as one.

Current resolution is to introduce id field that is composed of our calculated slug and leave slug as-is to avoid un-necessary changes.

Resolves: #58