Expose repository webhook as computed attribute
When code integrations are configured in Sleuth in read-only mode, it is
up to the user to manually add the relevant webhook to the relevant
repository. These webhooks require secret tokens, which are only
available in Sleuth after the code change source has been created.

This change allows a user of the Terraform provider to automatically
create the webhooks in their code integration source. For example, with
GitLab, it is possible to do:

  resource "sleuth_code_change_source" "example" {
    # [...]
  }

  resource "gitlab_project_hook" "example" {
    project = "${sleuth_code_change_source.example.repository.owner}/${sleuth_code_change_source.example.repository.name}"
    url     = sleuth_code_change_source.example.repository.webhook.url
    token   = sleuth_code_change_source.example.repository.webhook.secret

    job_events            = true
    merge_requests_events = true
    push_events           = true
    tag_push_events       = true
  }