Change struct to Object in resource model type
Terraform isn't able to deal with 'unknown' types in struct pointers,
and doesn't seem fully capable of representing null.

The recommendation in this case is to use an 'object' type:
https://developer.hashicorp.com/terraform/plugin/framework/handling-data/types/object

This change updates the webhook model to use this type instead, which
fixes a 'Value Conversion Error' outputted by Terraform when applying
certain changes that affect the webhook structure.