Fix panic when code deployment doesn't exist
When a code change source managed by Terraform and tracked in Terraform
state is deleted externally (e.g. by manually deleting it in the
interface), the Terraform provider currently panics upon trying to
apply without any Terraform changes.

This is because the `nil` return value in GetCodeChangeSource is never
checked: it is assumed that a nil error means a non-nil CodeChangeSource
pointer, which is a generally reasonable assumption to make in idiomatic
Go code [1].

Thus, this commit introduces a new error, which is returned when the
code change source cannot be found. This error is already being checked
in the one place that GetCodeChangeSource is invoked, and hence this
change results in a friendly, non-fatal error to the user, rather than a
runtime panic on a nil pointer dereference.

[1]: https://github.com/Antonboom/nilnil